+++ src/components/toasts/error-message.tsx
@@ -1,0 +1,14 @@
+interface Props {
+  message: string
+  label: string
+}
+
+export const ErrorMessage = (props: Props) => {
+  const { message, label } = props
+  return (
+    <div>
+      <span>{label}</span>
+      <p className="text-red-600">{message}</p>
+    </div>
+  )
+}
