+++ src/pages/_document.tsx
@@ -1,8 +1,12 @@
-import { Html, Main, NextScript } from 'next/document'
+import { Html, Head, Main, NextScript } from 'next/document'
+
 
 export default function Document() {
   return (
     <Html lang="en">
+      <Head>
+      
+      </Head>
       <body className="bg-slate-50 box-border">
         <Main />
         <NextScript />
